pyproject.toml
[project]
name = "nova-2-omni-demo"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.21",
    "botocore>=1.42.21",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "langchain>=1.2.0",
    "langchain-aws>=1.2.0",
    "langchain-core>=1.2.6",
    "langgraph>=1.0.5",
    "opencv-python>=4.12.0.88",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "requests>=2.32.5",
    "strands-agents>=1.21.0",
    "streamlit>=1.52.2",
]
